requirements.txt
streamlit>=1.30.0
newspaper3k>=0.2.8
beautifulsoup4>=4.12.0
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-google-genai>=1.0.0
langchain-groq>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
lxml_html_clean>=0.1.0
